Lint cleanup in aries_vcx
