+++ agents/rust/aries-vcx-agent/src/services/issuer.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use aries_vcx::{
-    core::profile::{profile::Profile, vdrtools_profile::VdrtoolsProfile},
+    core::profile::{vdrtools_profile::VdrtoolsProfile, Profile},
     handlers::{issuance::issuer::Issuer, util::OfferInfo},
     messages::{
         msg_fields::protocols::cred_issuance::v1::{
