+++ aries_vcx/src/core/profile/mod.rs
@@ -1,22 +1,31 @@
 pub mod ledger;
 #[cfg(feature = "modular_libs")]
 pub mod modular_libs_profile;
-pub mod profile;
 #[cfg(feature = "vdr_proxy_ledger")]
 pub mod vdr_proxy_profile;
 #[cfg(feature = "vdrtools")]
 pub mod vdrtools_profile;
 
-const DEFAULT_AML_LABEL: &str = "eula";
-
 use std::sync::Arc;
 
-use aries_vcx_core::ledger::{
-    base_ledger::{IndyLedgerRead, TxnAuthrAgrmtOptions},
-    indy_vdr_ledger::GetTxnAuthorAgreementData,
+#[cfg(feature = "migration")]
+use aries_vcx_core::WalletHandle;
+use aries_vcx_core::{
+    anoncreds::base_anoncreds::BaseAnonCreds,
+    ledger::{
+        base_ledger::{
+            AnoncredsLedgerRead, AnoncredsLedgerWrite, IndyLedgerRead, IndyLedgerWrite,
+            TxnAuthrAgrmtOptions,
+        },
+        indy_vdr_ledger::GetTxnAuthorAgreementData,
+    },
+    wallet::base_wallet::BaseWallet,
 };
+use async_trait::async_trait;
 
 use crate::errors::error::VcxResult;
+
+const DEFAULT_AML_LABEL: &str = "eula";
 
 pub async fn prepare_taa_options(
     ledger_read: Arc<dyn IndyLedgerRead>,
@@ -31,4 +40,27 @@ pub async fn prepare_taa_options(
     } else {
         Ok(None)
     }
+}
+
+#[async_trait]
+pub trait Profile: std::fmt::Debug + Send + Sync {
+    type LedgerRead: IndyLedgerRead + AnoncredsLedgerRead;
+    type LedgerWrite: IndyLedgerWrite + AnoncredsLedgerWrite;
+    type Anoncreds: BaseAnonCreds;
+    type Wallet: BaseWallet;
+
+    fn ledger_read(&self) -> &Self::LedgerRead;
+
+    fn ledger_write(&self) -> &Self::LedgerWrite;
+
+    fn anoncreds(&self) -> &Self::Anoncreds;
+
+    fn wallet(&self) -> &Self::Wallet;
+
+    #[cfg(feature = "migration")]
+    fn wallet_handle(&self) -> Option<WalletHandle> {
+        None
+    }
+
+    fn update_taa_configuration(&self, taa_options: TxnAuthrAgrmtOptions) -> VcxResult<()>;
 }
