+++ aries_vcx/src/protocols/issuance/holder/states/initial.rs
@@ -1,13 +1,9 @@
 use crate::errors::error::prelude::*;
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct InitialHolderState {}
+pub struct InitialHolderState;
 
 impl InitialHolderState {
-    pub fn new() -> Self {
-        Self {}
-    }
-
     pub fn is_revokable(&self) -> VcxResult<bool> {
         Err(AriesVcxError::from_msg(
             AriesVcxErrorKind::InvalidState,
