+++ aries_vcx/src/protocols/revocation_notification/sender/states/initial.rs
@@ -1,8 +1,2 @@
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub struct InitialState {}
-
-impl InitialState {
-    pub fn new() -> Self {
-        Self {}
-    }
-}
+pub struct InitialState;
