+++ aries_vcx/src/utils/mockdata/profile/mock_profile.rs
@@ -2,7 +2,7 @@ use aries_vcx_core::{ledger::base_ledger::TxnAuthrAgrmtOptions, wallet::mock_wal
 use async_trait::async_trait;
 
 use super::{mock_anoncreds::MockAnoncreds, mock_ledger::MockLedger};
-use crate::{core::profile::profile::Profile, errors::error::VcxResult};
+use crate::{core::profile::Profile, errors::error::VcxResult};
 
 /// Implementation of a [Profile] which uses [MockLedger], [MockAnoncreds] and [MockWallet] to
 /// return mock data for all Profile methods. Only for unit testing purposes
