+++ aries_vcx/tests/test_pool.rs
@@ -29,7 +29,7 @@ use aries_vcx::{
             create_and_write_test_schema,
         },
     },
-    core::profile::profile::Profile,
+    core::profile::Profile,
     errors::error::AriesVcxErrorKind,
     run_setup,
     utils::{
